NXT-100 Move chapter objects path calculation out of PublicationChapterMenu component
